[HatoholArmPluginBase] Fix a compile time warning on 32bit environment
HatoholArmPluginBase.cc: In member function 'virtual void HatoholArmPluginBase::getTimeOfLastEvent(const TriggerIdType&)::Callback::onGotReply(mlpl::SmartBuffer&, const HapiCommandHeader&)':
HatoholArmPluginBase.cc:263:4: warning: narrowing conversion of 'time->HapiResTimeOfLastEvent::nsec' from 'const uint32_t {aka const unsigned int}' to '__syscall_slong_t {aka long int}' inside { } is ill-formed in C++11 [-Wnarrowing]

+++ server/hap/HatoholArmPluginBase.cc
@@ -257,7 +257,10 @@ SmartTime HatoholArmPluginBase::getTimeOfLastEvent(
 			const HapiResTimeOfLastEvent *time =
 			  getObject()->getResponseBody
 			    <HapiResTimeOfLastEvent>(replyBuf);
-			timespec ts = {(time_t)time->sec, time->nsec};
+			timespec ts = {
+				(time_t)time->sec,
+				(long)time->nsec
+			};
 			lastTime = SmartTime(ts);
 			setSucceeded();
 		}
